:boom:  feat: [BREAKING] change the timing of collecting definition identifiers and footnote definition identifiers
Provide new match-block phase API registerDefinitionIdentifier and registerFootnoteDefinitionIdentifier
to register a definition identifier declaration and footnote definition identifier declaration respectly.

This change only affects users who implement a custom Definition or FootnoteDefinition tokenizer.

This change is to fix the bug that the reference link or footnote cannot be recognized when call the
parsePhrasingContent API in the parse-block phase.